backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.13
django-admin==2.0.2
django-excel-response2==3.0.6
django-filter==24.2
django-six==1.0.5
djangorestframework==3.15.2
excel-base==1.0.4
idna==3.7
isoweek==1.3.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
screen==1.0.1
six==1.16.0
sqlparse==0.5.0
TimeConvert==3.0.13
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
xlwt==1.3.0
